requirements.txt
playwright==1.41.0
python-dotenv==1.0.1
